+++ Fairing/Runge.h
@@ -1,4 +1,4 @@
 #pragma once
 #include "status.h"
-void runge(status &sv, double h);
+void runge(status &sv, long double h);
 void euler(status &sv, double h);
